+++ NEWS.md
@@ -29,3 +29,8 @@
 
 # RPEIF 1.0.8
 * Update to vignette with new Appendix B.
+
+# RPEIF 1.0.9
+* Rename Rachev ratio function.
+* Add the Downside Sharpe Ratio function.
+* Created utility functions for repetitive tasks.